requirements.txt
# Trading Bot Dependencies

# Web Framework
flask==3.0.0
werkzeug==3.0.0

# HTTP Requests
requests==2.31.0
urllib3==2.0.7

# WebSocket (SignalR)
signalrcore==0.9.5

# Timezone Support
pytz==2023.3
tzdata==2024.1

# Environment Variables
python-dotenv==1.0.0